+++ src/random.jl
@@ -17,7 +17,11 @@ function rng(init=false)
 end
 end
 
-# our version of srand sets both gpu and cpu
+"""
+    setseed(n::Integer)
+
+Run srand(n) on both cpu and gpu.
+"""
 function setseed(n::Integer)
     # need to regenerate RNG for the seed to take effect for some reason
     @cuda(curand,curandSetPseudoRandomGeneratorSeed,(Cptr,Culonglong),rng(true),n)