+++ src/reduction.jl
@@ -132,3 +132,4 @@ end
 
 # @primitive xentloss(x,p,d...),dy,y  (dy.*xentback(x,p,d...))
 
+Base.mean(a::KnetArray)=sum(a)/length(a)